normal order 1

+++ src/main/java/lk/ijse/company/model/OrderDetail.java
@@ -1,10 +1,7 @@
 package lk.ijse.company.model;
 
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 //@AllArgsConstructor
 //@NoArgsConstructor
@@ -40,7 +37,7 @@ public class OrderDetail implements Serializable {
         this.qty = qty;
     }
 
-    public double getUnitPrice() {
+    public BigDecimal getUnitPrice() {
         return unitPrice;
     }
